+++ user/test/client.go
@@ -29,6 +29,11 @@ type GetUserPermissionsOutput struct {
 }
 
 type Client struct {
+	EnsureAuthorizedInvocations        int
+	EnsureAuthorizedInputs             []context.Context
+	EnsureAuthorizedStub               func(ctx context.Context) error
+	EnsureAuthorizedOutputs            []error
+	EnsureAuthorizedOutput             *error
 	EnsureAuthorizedServiceInvocations int
 	EnsureAuthorizedServiceInputs      []context.Context
 	EnsureAuthorizedServiceStub        func(ctx context.Context) error
@@ -50,65 +55,85 @@ func NewClient() *Client {
 	return &Client{}
 }
 
-func (r *Client) EnsureAuthorizedService(ctx context.Context) error {
-	r.EnsureAuthorizedServiceInvocations++
-	r.EnsureAuthorizedServiceInputs = append(r.EnsureAuthorizedServiceInputs, ctx)
-	if r.EnsureAuthorizedServiceStub != nil {
-		return r.EnsureAuthorizedServiceStub(ctx)
+func (c *Client) EnsureAuthorized(ctx context.Context) error {
+	c.EnsureAuthorizedInvocations++
+	c.EnsureAuthorizedInputs = append(c.EnsureAuthorizedInputs, ctx)
+	if c.EnsureAuthorizedStub != nil {
+		return c.EnsureAuthorizedStub(ctx)
 	}
-	if len(r.EnsureAuthorizedServiceOutputs) > 0 {
-		output := r.EnsureAuthorizedServiceOutputs[0]
-		r.EnsureAuthorizedServiceOutputs = r.EnsureAuthorizedServiceOutputs[1:]
+	if len(c.EnsureAuthorizedOutputs) > 0 {
+		output := c.EnsureAuthorizedOutputs[0]
+		c.EnsureAuthorizedOutputs = c.EnsureAuthorizedOutputs[1:]
 		return output
 	}
-	if r.EnsureAuthorizedServiceOutput != nil {
-		return *r.EnsureAuthorizedServiceOutput
+	if c.EnsureAuthorizedOutput != nil {
+		return *c.EnsureAuthorizedOutput
+	}
+	panic("EnsureAuthorized has no output")
+}
+
+func (c *Client) EnsureAuthorizedService(ctx context.Context) error {
+	c.EnsureAuthorizedServiceInvocations++
+	c.EnsureAuthorizedServiceInputs = append(c.EnsureAuthorizedServiceInputs, ctx)
+	if c.EnsureAuthorizedServiceStub != nil {
+		return c.EnsureAuthorizedServiceStub(ctx)
+	}
+	if len(c.EnsureAuthorizedServiceOutputs) > 0 {
+		output := c.EnsureAuthorizedServiceOutputs[0]
+		c.EnsureAuthorizedServiceOutputs = c.EnsureAuthorizedServiceOutputs[1:]
+		return output
+	}
+	if c.EnsureAuthorizedServiceOutput != nil {
+		return *c.EnsureAuthorizedServiceOutput
 	}
 	panic("EnsureAuthorizedService has no output")
 }
 
-func (r *Client) EnsureAuthorizedUser(ctx context.Context, targetUserID string, permission string) (string, error) {
-	r.EnsureAuthorizedUserInvocations++
-	r.EnsureAuthorizedUserInputs = append(r.EnsureAuthorizedUserInputs, EnsureAuthorizedUserInput{Context: ctx, TargetUserID: targetUserID, Permission: permission})
-	if r.EnsureAuthorizedUserStub != nil {
-		return r.EnsureAuthorizedUserStub(ctx, targetUserID, permission)
+func (c *Client) EnsureAuthorizedUser(ctx context.Context, targetUserID string, permission string) (string, error) {
+	c.EnsureAuthorizedUserInvocations++
+	c.EnsureAuthorizedUserInputs = append(c.EnsureAuthorizedUserInputs, EnsureAuthorizedUserInput{Context: ctx, TargetUserID: targetUserID, Permission: permission})
+	if c.EnsureAuthorizedUserStub != nil {
+		return c.EnsureAuthorizedUserStub(ctx, targetUserID, permission)
 	}
-	if len(r.EnsureAuthorizedUserOutputs) > 0 {
-		output := r.EnsureAuthorizedUserOutputs[0]
-		r.EnsureAuthorizedUserOutputs = r.EnsureAuthorizedUserOutputs[1:]
+	if len(c.EnsureAuthorizedUserOutputs) > 0 {
+		output := c.EnsureAuthorizedUserOutputs[0]
+		c.EnsureAuthorizedUserOutputs = c.EnsureAuthorizedUserOutputs[1:]
 		return output.AuthorizedUserID, output.Error
 	}
-	if r.EnsureAuthorizedUserOutput != nil {
-		return r.EnsureAuthorizedUserOutput.AuthorizedUserID, r.EnsureAuthorizedUserOutput.Error
+	if c.EnsureAuthorizedUserOutput != nil {
+		return c.EnsureAuthorizedUserOutput.AuthorizedUserID, c.EnsureAuthorizedUserOutput.Error
 	}
 	panic("EnsureAuthorizedUser has no output")
 }
 
-func (r *Client) GetUserPermissions(ctx context.Context, requestUserID string, targetUserID string) (user.Permissions, error) {
-	r.GetUserPermissionsInvocations++
-	r.GetUserPermissionsInputs = append(r.GetUserPermissionsInputs, GetUserPermissionsInput{Context: ctx, RequestUserID: requestUserID, TargetUserID: targetUserID})
-	if r.GetUserPermissionsStub != nil {
-		return r.GetUserPermissionsStub(ctx, requestUserID, targetUserID)
+func (c *Client) GetUserPermissions(ctx context.Context, requestUserID string, targetUserID string) (user.Permissions, error) {
+	c.GetUserPermissionsInvocations++
+	c.GetUserPermissionsInputs = append(c.GetUserPermissionsInputs, GetUserPermissionsInput{Context: ctx, RequestUserID: requestUserID, TargetUserID: targetUserID})
+	if c.GetUserPermissionsStub != nil {
+		return c.GetUserPermissionsStub(ctx, requestUserID, targetUserID)
 	}
-	if len(r.GetUserPermissionsOutputs) > 0 {
-		output := r.GetUserPermissionsOutputs[0]
-		r.GetUserPermissionsOutputs = r.GetUserPermissionsOutputs[1:]
+	if len(c.GetUserPermissionsOutputs) > 0 {
+		output := c.GetUserPermissionsOutputs[0]
+		c.GetUserPermissionsOutputs = c.GetUserPermissionsOutputs[1:]
 		return output.Permissions, output.Error
 	}
-	if r.GetUserPermissionsOutput != nil {
-		return r.GetUserPermissionsOutput.Permissions, r.GetUserPermissionsOutput.Error
+	if c.GetUserPermissionsOutput != nil {
+		return c.GetUserPermissionsOutput.Permissions, c.GetUserPermissionsOutput.Error
 	}
 	panic("GetUserPermissions has no output")
 }
 
-func (r *Client) AssertOutputsEmpty() {
-	if len(r.EnsureAuthorizedServiceOutputs) > 0 {
+func (c *Client) AssertOutputsEmpty() {
+	if len(c.EnsureAuthorizedOutputs) > 0 {
+		panic("EnsureAuthorizedOutputs is not empty")
+	}
+	if len(c.EnsureAuthorizedServiceOutputs) > 0 {
 		panic("EnsureAuthorizedServiceOutputs is not empty")
 	}
-	if len(r.EnsureAuthorizedUserOutputs) > 0 {
+	if len(c.EnsureAuthorizedUserOutputs) > 0 {
 		panic("EnsureAuthorizedUserOutputs is not empty")
 	}
-	if len(r.GetUserPermissionsOutputs) > 0 {
+	if len(c.GetUserPermissionsOutputs) > 0 {
 		panic("GetUserPermissionsOutputs is not empty")
 	}
 }